Implemented post download

+++ app/src/main/java/com/example/connectify/core/presentation/components/StandardBottomSheet.kt
@@ -1,5 +1,6 @@
 package com.example.connectify.core.presentation.components
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -18,7 +19,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -29,7 +29,6 @@ import com.example.connectify.R
 import com.example.connectify.core.presentation.ui.theme.HintGray
 import com.example.connectify.core.presentation.ui.theme.IconSizeMedium
 import com.example.connectify.core.presentation.ui.theme.IconSizeSmall
-import com.example.connectify.core.presentation.ui.theme.SpaceLarge
 import com.example.connectify.core.presentation.ui.theme.SpaceMedium
 import com.example.connectify.core.presentation.ui.theme.SpaceSmall
 
@@ -38,7 +37,10 @@ import com.example.connectify.core.presentation.ui.theme.SpaceSmall
 fun StandardBottomSheet(
     title: String,
     bottomSheetState: SheetState,
+    showDownloadOption: Boolean = false,
+    showDeleteOption: Boolean = false,
     onDismissRequest: () -> Unit = {},
+    onDownloadClick: () -> Unit = {},
     onDeleteClick: () -> Unit = {},
     onCancelClick: () -> Unit = {}
 ) {
@@ -74,32 +76,62 @@ fun StandardBottomSheet(
                 thickness = 0.2.dp,
                 color = HintGray
             )
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .size(
-                        height = 50.dp,
-                        width = 100.dp
+            if(showDownloadOption) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .size(
+                            height = 50.dp,
+                            width = 100.dp
+                        )
+                        .clickable {
+                            onDownloadClick()
+                        },
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Spacer(modifier = Modifier.width(SpaceMedium))
+                    Icon(
+                        painter = painterResource(id = R.drawable.download_icon),
+                        contentDescription = null,
+                        modifier = Modifier.size(IconSizeSmall),
                     )
-                    .clickable {
-                        onDeleteClick()
-                    },
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Spacer(modifier = Modifier.width(SpaceMedium))
-                Icon(
-                    painter = painterResource(id = R.drawable.delete_icon),
-                    contentDescription = null,
-                    modifier = Modifier.size(IconSizeSmall),
-                    tint = Color.Red
-                )
-                Spacer(modifier = Modifier.width(SpaceSmall))
-                Text(
-                    text = stringResource(id = R.string.delete),
-                    style = MaterialTheme.typography.titleSmall.copy(
-                        fontSize = 18.sp
+                    Spacer(modifier = Modifier.width(SpaceSmall))
+                    Text(
+                        text = stringResource(id = R.string.download),
+                        style = MaterialTheme.typography.labelSmall.copy(
+                            fontSize = 18.sp
+                        )
                     )
-                )
+                }
+            }
+
+            if(showDeleteOption) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .size(
+                            height = 50.dp,
+                            width = 100.dp
+                        )
+                        .clickable {
+                            onDeleteClick()
+                        },
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Spacer(modifier = Modifier.width(SpaceMedium))
+                    Image(
+                        painter = painterResource(id = R.drawable.delete_icon),
+                        contentDescription = null,
+                        modifier = Modifier.size(IconSizeSmall)
+                    )
+                    Spacer(modifier = Modifier.width(SpaceSmall))
+                    Text(
+                        text = stringResource(id = R.string.delete),
+                        style = MaterialTheme.typography.titleSmall.copy(
+                            fontSize = 18.sp
+                        )
+                    )
+                }
             }
             Row(
                 modifier = Modifier
@@ -117,9 +149,9 @@ fun StandardBottomSheet(
                 Icon(
                     painter = painterResource(id = R.drawable.cancel_icon),
                     contentDescription = null,
-                    modifier = Modifier.size(24.dp)
+                    modifier = Modifier.size(IconSizeSmall)
                 )
-                Spacer(modifier = Modifier.width(5.dp))
+                Spacer(modifier = Modifier.width(SpaceSmall))
                 Text(
                     text = stringResource(id = R.string.cancel),
                     style = MaterialTheme.typography.labelSmall.copy(
